♻️ refactor : 오타 수정 및 리팩토링
1. 오타
- ResponseProductHistroyDTO.java 필드명 수정 create_at -> created_at
- ProductHistory.java 필드명 수정 create_at -> created_at
- ResponseProductMasterDetailDTO.java 필드명 수정 reatalProducts -> rentalProducts

2. 리팩토링
- ProductHistoryRepository 키값 타입 변경 Integer -> Long
- ProductRepositoryImpl.java 에서 productCd 조건을 매개변수로 비교가 아닌 entity로 비교하고 있던걸 매개변수랑 비교 하도록 수정
- 제품 마스터 신규 정보 등록시 반환 status 를 201로 반환 하도록 수정

+++ src/main/java/org/ateam/oncare/careproduct/command/repository/ProductHistoryRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProductHistoryRepository extends JpaRepository<ProductHistory,Integer>, ProductHistoryRepositoryCustom {
+public interface ProductHistoryRepository extends JpaRepository<ProductHistory,Long>, ProductHistoryRepositoryCustom {
 }
